+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.6.0
 
 - Set environment tag for native layer.
 - Actually fallback to current stack trace if the one passed to